[core] Leverage ID field in TTL extension (100x+ speed increase)
To further increase speed, the expiration time is encoded in the
ttl-drawer document ID.  Searching for expired documents is thus
searching for IDs 'smaller than'. This takes advantage of the fact
that the ID (the primary key) is the most likely field to be
optimized. With the Codax implementation, this optimization lowers
the time taken to search 20k documents from ~500ms to <5ms.